Program añadido y fixes para algunas clases.

+++ src/Library/Elemento.cs
@@ -7,10 +7,10 @@ namespace Library
     {
         public int Ataque{get; private set;}
         public int Defensa{get; private set;}
-        public int Potenciador{get; private set;}
+        public double Potenciador{get; private set;}
         public string Nombre{get; private set;}
 
-        public Elemento(string pNombre,int pPotenciador,int pAtaque, int pDefensa)
+        public Elemento(string pNombre,double pPotenciador,int pAtaque, int pDefensa)
         {
             this.Ataque=pAtaque;
             this.Defensa=pDefensa;
